Add support for dispatch, bind.
Dispatch just needs to run on the main thread. We can't pass a Go
pointer to a struture containing Go pointers to C, so to pass functions
to the main thread for dispatch, functions are put on a queue.

Bind required more work. It is intended to be compatible with
webview/webview but internally works a lot differently.